Break up logging
