+++ neuro_san/service/http/server/http_server_app.py
@@ -23,7 +23,7 @@ from threading import Thread
 from tornado.web import Application
 from tornado.ioloop import IOLoop
 
-from neuro_san.service.logging.event_loop_logger import EventLoopLogger
+from neuro_san.service.interfaces.event_loop_logger import EventLoopLogger
 
 
 class HttpServerApp(Application):
