+++ neuro_san/service/http/server/http_sidecar.py
@@ -25,6 +25,7 @@ from tornado.ioloop import IOLoop
 from neuro_san.interfaces.concierge_session import ConciergeSession
 from neuro_san.internals.network_providers.service_agent_network_storage import ServiceAgentNetworkStorage
 from neuro_san.internals.network_providers.single_agent_network_provider import SingleAgentNetworkProvider
+from neuro_san.service.generic.agent_server_logging import AgentServerLogging
 from neuro_san.service.generic.async_agent_service import AsyncAgentService
 from neuro_san.service.http.handlers.health_check_handler import HealthCheckHandler
 from neuro_san.service.http.handlers.connectivity_handler import ConnectivityHandler
@@ -37,8 +38,7 @@ from neuro_san.service.http.interfaces.agents_updater import AgentsUpdater
 from neuro_san.service.http.logging.http_logger import HttpLogger
 from neuro_san.service.http.server.http_server_app import HttpServerApp
 from neuro_san.service.interfaces.agent_server import AgentServer
-from neuro_san.service.logging.agent_server_logging import AgentServerLogging
-from neuro_san.service.logging.event_loop_logger import EventLoopLogger
+from neuro_san.service.interfaces.event_loop_logger import EventLoopLogger
 from neuro_san.session.direct_concierge_session import DirectConciergeSession
